optic/requirements.txt
pytest
flask
mypy-protobuf
absl-py